Task_1_3_1

+++ Task_1_3_1/src/test/java/ru/nsu/demidov/substring/SubStringFinderTest.java
@@ -1,12 +1,10 @@
 package ru.nsu.demidov.substring;
 
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
+import java.io.*;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -23,12 +21,15 @@ public class SubStringFinderTest {
     @Test
     public void bigDataTesting() throws Exception {
         long OCCURENCES = 2000000;
-        List<Integer> result = SubStringFinder.myFind("input.txt", "boobies");
+        try (BufferedWriter writer = new BufferedWriter(new FileWriter(tempFile))) {
+            StringBuilder sb = new StringBuilder();
+            for (int i = 0; i < OCCURENCES; i++) {
+                sb.append("boobies");
+            }
+            writer.write(sb.toString());
+        }
+        List<Integer> result = SubStringFinder.myFind(new FileInputStream(tempFile), "boobies");
         assertEquals(OCCURENCES, result.size());
-        /*for (int i = 0; i < result.size(); i++) {
-            assertEquals(i * 19 + 7, result.get(i).intValue());
-        }*/
-
     }
 
     @Test
@@ -36,10 +37,8 @@ public class SubStringFinderTest {
         long OCCURENCES = 10000000L;
         try (BufferedWriter writer = new BufferedWriter(new FileWriter(tempFile))) {
             StringBuilder sb = new StringBuilder();
-            for (long jL = 0L; jL < OCCURENCES; jL++) {
-                for (long iL = 0L; iL < OCCURENCES; iL++) {
-                    sb.append("boobsieboobiesboobs");
-                }
+            for (long iL = 0L; iL < OCCURENCES; iL++) {
+                sb.append("boobsieboobiesboobs");
             }
             sb.append("bebra");
             writer.write(sb.toString());
@@ -47,9 +46,8 @@ public class SubStringFinderTest {
             System.out.println(exception.getMessage());
         }
         String subString = "bebra";
-        List<Integer> result = SubStringFinder.myFind(tempFile.getAbsolutePath(), subString);
+        List<Integer> result = SubStringFinder.myFind(new FileInputStream(tempFile), subString);
         assertEquals(1, result.size());
-        tempFile.delete();
     }
 
     @Test
@@ -65,11 +63,17 @@ public class SubStringFinderTest {
             System.out.println(exception.getMessage());
         }
         String subString = "ааааааааааа";
-        List<Integer> result = SubStringFinder.myFind(tempFile.getAbsolutePath(), subString);
+        List<Integer> result = SubStringFinder.myFind(new FileInputStream(tempFile), subString);
         assertEquals(OCCURENCES * subString.length() * 2, result.size());
         for (int i = 0; i < result.size(); i++) {
             assertEquals(i, result.get(i).intValue());
         }
-        tempFile.delete();
+    }
+
+    @AfterEach
+    public void deleteEach() {
+        if (tempFile != null && tempFile.exists()) {
+            tempFile.delete();
+        }
     }
 }
